fix: Move and fix import Avatar
