+++ src/components/index.tsx
@@ -3,3 +3,4 @@ export { ViewHeader } from './ViewHeader/ViewHeader'
 export { AppHeader } from './AppHeader'
 export { ViewWrapper } from './ViewWrapper'
 export { Awareness } from './Awareness'
+export { Avatar } from './Avatar'
